requirements.txt
sqlalchemy
passlib
fastapi
uvicorn
bcrypt
python-jose
python-multipart
aiofiles
